[refactor] Organize routes and schemas into subfolders
