+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@prisma/client": "^6.13.0",
     "dotenv": "^17.2.1",
-    "express": "^5.1.0"
+    "express": "^5.1.0",
+    "zod": "^4.0.15"
   }
 }
